getHilfeSolution/modules/getHilfeModule/requirements.txt
azure-iot-device
psycopg2-binary
datetime